+++ src/kiota/Handlers/KiotaSeachBasedCommandHandler.cs
@@ -1,7 +1,4 @@
-﻿using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.Extensions.Logging;
+﻿using Microsoft.Extensions.Logging;
 
 namespace kiota.Handlers;